app/requirements.txt
fastapi
pdfminer.six
uvicorn
python-dotenv
python-multipart
pytest
httpx
pytesseract
PyMuPDF
numpy
opencv-python